feat: make key required for DragAndDropItem and DragAndDropList
Previously, the reoder logic uses compare instance. Comparing using
instance is not reliable because the item is rebuilt, hence making the
compare sometimes fail. This might be a bug from the orignal author.

This commit fixed this by requiring the item widget to have key. We then
use key for all comparisons.